requirements.txt
requests==2.32.3
feedparser==6.0.11
python-dotenv==1.0.1
pytest==8.3.4
